Rule trigger notification updates.
When a template is empty, it will no longer send the notification to that
role or user.

Added extra configuration item to specify the noreply emails full name.

The user from email field is now the noreply address. The name is either
from the system when it is sent to a user, or from the user when it is sent
to a role.