Modifications to the state change history table and views.
A new field has been added to the DB for local_extension_hist_state to track the length of the extension request at that date.
A new output function in the state class render_state_history() will now display all state changes and as many lengths as possible.
When new requests are made, an initial state is added to the history table.
The version has been updated to implement these DB changes.